experiments/workloads/bench1/src/requirements.txt
flask
waitress
python-socketio[asyncio_client]
nest_asyncio